Tidy up bugs and remove dependencies
