+++ mbuild/__init__.py
@@ -6,6 +6,5 @@ from mbuild.pattern import *
 from mbuild.packing import *
 from mbuild.port import Port
 from mbuild.recipes import *
-from mbuild.formats import *
 
 from mbuild.version import version
