+++ mbuild/lib/surfaces/betacristobalite.py
@@ -35,6 +35,9 @@ class Betacristobalite(mb.Compound):
                 mb.rotate_around_x(port, np.pi/2)
                 mb.translate(port, particle.pos + np.array([0, 0, .1]))
                 self.add(port, 'port_{}'.format(count))
+                particle.name = 'O'  # Strip numbers required in .mol2 files.
+            elif particle.name.startswith('Si'):
+                particle.name = 'Si'
 
 if __name__ == "__main__":
     single = Betacristobalite()
