+++ mbuild/tests/test_pattern.py
@@ -16,3 +16,27 @@ class TestPattern(BaseTest):
             pattern = mb.Random2DPattern(101)
             chains, backfills = pattern.apply_to_compound(
                 guest=alkyl, host=betacristobalite, backfill=ch3)
+
+    def test_random_2d(self):
+        pattern = mb.Random2DPattern(100)
+        assert len(pattern) == 100
+
+    def test_random_3d(self):
+        pattern = mb.Random3DPattern(100)
+        assert len(pattern) == 100
+
+    def test_grid_2d(self):
+        pattern = mb.Grid2DPattern(10, 5)
+        assert len(pattern) == 50
+
+    def test_grid_3d(self):
+        pattern = mb.Grid3DPattern(10, 5, 2)
+        assert len(pattern) == 100
+
+    def test_sphere(self):
+        pattern = mb.SpherePattern(100)
+        assert len(pattern) == 100
+
+    def test_disk(self):
+        pattern = mb.DiskPattern(100)
+        assert len(pattern) == 100
